Move deployment_to_schema and schema_to_deployment to deployment.py
