+++ python/ray/serve/api.py
@@ -50,7 +50,11 @@ from ray.serve.constants import (
     DEFAULT_HTTP_PORT,
 )
 from ray.serve.controller import ServeController
-from ray.serve.deployment import Deployment
+from ray.serve.deployment import (
+    Deployment,
+    deployment_to_schema,
+    schema_to_deployment,
+)
 from ray.serve.exceptions import RayServeException
 from ray.serve.generated.serve_pb2 import (
     DeploymentRoute,
@@ -66,7 +70,6 @@ from ray.serve.utils import (
     format_actor_name,
     get_current_node_resource_key,
     get_random_letters,
-    get_deployment_import_path,
     in_interactive_shell,
     DEFAULT,
 )
@@ -74,8 +77,6 @@ from ray.util.annotations import PublicAPI
 import ray
 from ray import cloudpickle
 from ray.serve.schema import (
-    RayActorOptionsSchema,
-    DeploymentSchema,
     DeploymentStatusSchema,
     ServeApplicationSchema,
     ServeApplicationStatusSchema,
@@ -1368,74 +1369,6 @@ def build(target: Union[DeploymentNode, DeploymentFunctionNode]) -> Application:
     # TODO(edoakes): this should accept host and port, but we don't
     # currently support them in the REST API.
     return Application(pipeline_build(target))
-
-
-def deployment_to_schema(d: Deployment) -> DeploymentSchema:
-    """Converts a live deployment object to a corresponding structured schema.
-
-    If the deployment has a class or function, it will be attemptetd to be
-    converted to a valid corresponding import path.
-
-    init_args and init_kwargs must also be JSON-serializable or this call will
-    fail.
-    """
-    from ray.serve.pipeline.json_serde import convert_to_json_safe_obj
-
-    if d.ray_actor_options is not None:
-        ray_actor_options_schema = RayActorOptionsSchema.parse_obj(d.ray_actor_options)
-    else:
-        ray_actor_options_schema = None
-
-    return DeploymentSchema(
-        name=d.name,
-        import_path=get_deployment_import_path(
-            d, enforce_importable=True, replace_main=True
-        ),
-        init_args=convert_to_json_safe_obj(d.init_args, err_key="init_args"),
-        init_kwargs=convert_to_json_safe_obj(d.init_kwargs, err_key="init_kwargs"),
-        num_replicas=d.num_replicas,
-        route_prefix=d.route_prefix,
-        max_concurrent_queries=d.max_concurrent_queries,
-        user_config=d.user_config,
-        autoscaling_config=d._config.autoscaling_config,
-        graceful_shutdown_wait_loop_s=d._config.graceful_shutdown_wait_loop_s,
-        graceful_shutdown_timeout_s=d._config.graceful_shutdown_timeout_s,
-        health_check_period_s=d._config.health_check_period_s,
-        health_check_timeout_s=d._config.health_check_timeout_s,
-        ray_actor_options=ray_actor_options_schema,
-    )
-
-
-def schema_to_deployment(s: DeploymentSchema) -> Deployment:
-    from ray.serve.pipeline.json_serde import convert_from_json_safe_obj
-
-    if s.ray_actor_options is None:
-        ray_actor_options = None
-    else:
-        ray_actor_options = s.ray_actor_options.dict(exclude_unset=True)
-
-    config = DeploymentConfig.from_default(
-        ignore_none=True,
-        num_replicas=s.num_replicas,
-        user_config=s.user_config,
-        max_concurrent_queries=s.max_concurrent_queries,
-        _autoscaling_config=s.autoscaling_config,
-        _graceful_shutdown_wait_loop_s=s.graceful_shutdown_wait_loop_s,
-        _graceful_shutdown_timeout_s=s.graceful_shutdown_timeout_s,
-        _health_check_period_s=s.health_check_period_s,
-        _health_check_timeout_s=s.health_check_timeout_s,
-    )
-    
-    return Deployment(
-        func_or_class=s.import_path,
-        name=s.name,
-        config=config,
-        init_args=convert_from_json_safe_obj(s.init_args, err_key="init_args"),
-        init_kwargs=convert_from_json_safe_obj(s.init_kwargs, err_key="init_kwargs"),
-        route_prefix=s.route_prefix,
-        ray_actor_options=ray_actor_options,
-        _internal=True,
-    )
 
 
 def serve_application_to_schema(
